+++ src/main/java/org/patternfly/component/alert/AlertType.java
@@ -26,7 +26,6 @@ import static org.patternfly.layout.PredefinedIcon.infoCircle;
 
 public enum AlertType {
 
-    @Deprecated
     custom(bell, Status.custom, "custom alert"),
 
     info(infoCircle, Status.info, "info alert"),