+++ src/main/java/org/patternfly/component/brand/Brand.java
@@ -30,7 +30,8 @@ import static org.patternfly.layout.Classes.component;
  *
  * @see <a href= "https://www.patternfly.org/components/brand/html">https://www.patternfly.org/components/brand/html</a>
  */
-public class Brand extends BaseComponent<HTMLImageElement, Brand> implements UnderDevelopment {
+@UnderDevelopment
+public class Brand extends BaseComponent<HTMLImageElement, Brand> {
 
     // ------------------------------------------------------ factory methods